% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charcount.R
\name{charcount}
\alias{charcount}
\title{count the letter frequency of input string}
\usage{
charcount(instring)
}
\value{
returns a frequency table
}
\description{
count the letter frequency of input string
}
