% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypt.R
\name{crypt}
\alias{crypt}
\title{encrypt/decrypt a provided string using
the provided alphabet}
\usage{
crypt(instring, alphabet)
}
\value{
returns output_string
}
\description{
encrypt/decrypt a provided string using
the provided alphabet
}
