% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monogram_alpha.R
\name{monogram_alpha}
\alias{monogram_alpha}
\title{perform a frequency count of only alphabetic characters}
\usage{
monogram_alpha(inalphabet)
}
\value{
returns a frequency table
}
\description{
perform a frequency count of only alphabetic characters
}
