% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_alphabet.R
\name{shift_alphabet}
\alias{shift_alphabet}
\title{Perform a shift of a provided alphabet by the
number of provided characters}
\usage{
shift_alphabet(in_alphabet, shift)
}
\value{
returns the resulting alphabet
}
\description{
Perform a shift of a provided alphabet by the
number of provided characters
}
