% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_caesar.R
\name{solve_caesar}
\alias{solve_caesar}
\title{Perform a brute force search on all alphabet
shifts and test for possible solution based on
chi square results compared to a provided
alphabet frequency distribution}
\usage{
solve_caesar(ciphertext, expected_dist)
}
\value{
returns a dataframe of results
}
\description{
Perform a brute force search on all alphabet
shifts and test for possible solution based on
chi square results compared to a provided
alphabet frequency distribution
}
